requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
--extra-index-url https://download.pytorch.org/whl/cu113

antlr4-python3-runtime==4.9.3
    # via
    #   hydra-core
    #   omegaconf
beautifulsoup4==4.11.1
    # via pytransform3d
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
colorama==0.4.6
    # via tqdm
contourpy==1.0.5
    # via matplotlib
cycler==0.11.0
    # via matplotlib
fonttools==4.37.4
    # via matplotlib
hydra-core==1.2.0
    # via
    #   -r requirements.in
    #   hydra-zen
hydra-zen==0.8.0
    # via -r requirements.in
idna==3.4
    # via requests
kiwisolver==1.4.4
    # via matplotlib
lxml==4.9.1
    # via pytransform3d
matplotlib==3.6.1
    # via
    #   -r requirements.in
    #   pytransform3d
numpy==1.23.4
    # via
    #   -r requirements.in
    #   contourpy
    #   matplotlib
    #   pytransform3d
    #   scipy
    #   torchvision
omegaconf==2.2.3
    # via hydra-core
packaging==21.3
    # via
    #   hydra-core
    #   matplotlib
pillow==9.2.0
    # via
    #   matplotlib
    #   torchvision
pyparsing==3.0.9
    # via
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via matplotlib
pytransform3d==2.0.0
    # via -r requirements.in
pyyaml==6.0
    # via omegaconf
requests==2.28.1
    # via torchvision
scipy==1.9.3
    # via pytransform3d
six==1.16.0
    # via python-dateutil
soupsieve==2.3.2.post1
    # via beautifulsoup4
torch==1.12.1+cu113
    # via
    #   -r requirements.in
    #   torchvision
torchvision==0.13.1+cu113
    # via -r requirements.in
tqdm==4.64.1
    # via -r requirements.in
typing-extensions==4.4.0
    # via
    #   hydra-zen
    #   torch
    #   torchvision
urllib3==1.26.12
    # via requests
